// generated/src/app/api-models/system-code-type.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class SystemCodeType {
  systemCodeType: string;
  systemCodeTypeDesc: string;
  systemCodeTypeDataType: string;
  systemCodeTypeMaxLength: string;
  systemCodeTypeCase: string;
  systemCodeTypeUpd: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}